requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
pyyaml